Fix missing initializer warning in ext_authz tests on newer version of clang (#38177)
I'm trying to make Envoy build with newer version of LLVM toolchain,
specifically 18. I'm hitting a warning about missing initializers of the
Request structure in tests when dedicated initializers are used to set
values of some, but not all fields.

We can easily address this warning by just providing the default
initializers in the struct definition. That's what this change does.

Related to work in #37911

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>