Add DNS query Trace of NotStarted before queuing.
Before this change, it was possible that the `addTrace(NotStarted)`
occured after the resolver thread picked up the query and started
work. This can be seen when running the getaddrinfo tests 1000 times
in the form of failure around or just under 1% of the time.

bazel test --config=libc++ --runs_per_test 1000 //test/extensions/network/dns_resolver/getaddrinfo:getaddrinfo_test

Signed-off-by: Alejandro R. Sedeño <asedeno@google.com>