Fix MSAN warnings on clang-18 due to destruction order issues
This change contains multiple MSAN fixes detected by clang-18. They all
in different places, but all of them are caused by a similar issue.

In C++ members of the class are always created in the same order in
which they are declared in the class. And they are always destroyed in
reverse order. So the first memeber of the class is created first and
destroyed last.

When there are interdependencies between the members of a class the
order of destruction becomes important. We typically don't want to
access data of the already destroyed members in destructors of the
members that are being destroyed.

MSAN on clang-18 detected a bunch of cases where we access some fields
after they were destroyed. In all but one case, issues happens in tests
only and not in the Envoy binary. There is one case where a similar
problem occurs in regular Envoy code, but this issue is benign, because
we are accessing an atomic variable that does not trully get destroyed.

So, all-in-all, these are not serious issues, but we want to fix it to
make clang MSAN tests happy on new LLVM toolchain.

I also sharded a few tests, because it was the only way I could actually
run them through the end within given timeouts on my workstation.
Hopefully that's not a problem.

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>