logger: Create only one Context per process for mobile
The Logger::Context operates as a stack. When a new one is created, it
becomes the active Context. When it gets destroyed, the previous Context
is popped to become the activate one.

This can lead to lifetime bugs in Envoy Mobile. If multiple engines are
created, the destruction of those engines can cause indeterminate
destruction of the Logger::Context objects.

In this commit, the behavior is changed so that the Logger::Context
initialization in mobile is done one time only, outside of
StrippedMainBase. And StrippedMainBase will only initialize a
Logger::Context if one is not already active.

Signed-off-by: Ali Beyad <abeyad@google.com>