Add integration test for upstream filters with AsyncClient (#38068)
Commit Message: Add integration test for upstream filters with
AsyncClient
Additional Description: Behaviors were non-obvious to me and as far as I
can tell there were no tests ensuring that the behavior is what it is;
this adds a test for "AsyncClient bypasses router upstream filters but
uses cluster upstream filters" and a test for "if both router and
cluster upstream filters are set, router upstreams are ignored".
Risk Level: None, test-only.
Testing: Yes it is.
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

---------

Signed-off-by: Raven Black <ravenblack@dropbox.com>