Feature: Golang Http Tcp Bridge (#36667)
As mentioned in the proposal:
https://github.com/envoyproxy/envoy/issues/35749 , this PR is to support
using Golang to extend TCP upstream proxy, to make changes to
connections and data messages in the http2tcp situation of envoy.
(this PR does)

### Here is my thought about  Golang extension function points:
1. Support encoding message processing for upstream TCP requests (route
and cluster have been determined, and targeted message processing can be
performed for route and cluster)
2. Support the handling of conn connection status during the encoding
stage of upstream TCP requests (for example, by setting end_stream=false
to avoid envoy semi connected status)
3. Support decoding message processing and aggregation for upstream TCP
response in onUpstreamData.(for example, by setting end_stream=true to
indicate that the message is encapsulated and can be passed to
downstream)
4. Aggregate the tcp messages received multiple times by onUpstreamData
of tcp response.
5. Support obtaining route and cluster information, which can be
referenced for targeted processing in the above stages.

### What we can do with this  Golang extension:
With this golang extension, developers can quickly get started with
envoy and use golang to implement http2tcp such as http2dubbo、http2rpc.


Signed-off-by: duxin40 <duxin40@gamil.com>
Signed-off-by: duxin40 <33946910+duxin40@users.noreply.github.com>