ext_proc: least surprise behavior for DEFAULT in mode_override.
The mode_override feature has a very surprising behavior, in which unset fields in ProcessingMode,
revert behavior back to the default. For example, it is the default behavior that resposne headers
are sent, but if the filter is configured not to send response headers in processing_mode, and a
mode_override is sent to explicitly send request body, this then inadvertently triggers the sending
of response headers as well, as the mode has been overriden back to the "default". We have seen
ext_proc users of mode_override being affected by these confusing semantics.

It seems far more sensible that a mode_override only affects the procesisng mode features explicitly
set in the mode_override message. Technically this could be a breaking wire change, but it seems
unlikely that anyone is relying on the default override, as they would most likely just set response
headers to SEND instead of DEFAULT if that was the intended behavior.

This is a protocol change proposal, which will be augmented with implementation if there is
agreement on this approach.

Risk level: Medium
Testing: TODO

Signed-off-by: Harvey Tuch <htuch@google.com>