Explicitly use void in the C function list of arguments for functions that take no arguments
Using void in the list of function arguments in C++ is allowed for
functions that take no arguments, but it's rather uncommon.

However, when it comes to plain C, there is an actual difference between
a function declaration `int foo()` and `int foo(void)`. The first one
says that there is a function foo that returns an integer and nothing is
known about the arguments the function takes. While the second one
explicitly declares that the function takes no argument.

In practice this difference does not often matter, however, from the
language point of view, they are not the same and it triggers UBSAN when
running Envoy tests on clang-18.

Specifically, UBSAN complains that tests attempt to call a function
thought a pointer of incompatible function type.

I don't believe it causes any issues in practice for Envoy, but
addressing these warnings will help us to move to a newer version of
LLVM toolchains for Envoy.

NOTE: There are some alternative ways we could address this issue as
well, for example, we can change all the examples to use C++ instead of
C, and, I belive, it would also address the issue, but it's a bit more
invasive and I opted not to convert all the tests because of is
practically a false positive UBSAN complain.

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>