feature: inbound only graceful draining (#37873)
Fixes #35020. The inbound_only query param for the drain_listeners admin
endpoint doesn't work with the graceful query parameter. As a result,
outbound listeners will send connection: close headers to upstreams
which is undesired. This PR adds the ability for drain_manager to drain
in a single direction.

Prior to this change, the `inbound_only` query
param on the [`drain_listeners` admin endpoint
](https://www.envoyproxy.io/docs/envoy/latest/operations/admin#operations-admin-interface-drain)
only modified which listeners were stopped. If `graceful` is set,
listeners of all directions are drained, regardless if `inbound_only` is
set. If `skip_exit` is set, `inbound_only` has zero effect. This PR adds
the ability to drain only inbound listeners, allowing outbound listeners
to continue functioning as normal. This is useful in the Kubernetes
sidecar use-case where outbound traffic should not set connection: close
headers to the upstream.

Risk Level: Medium
fixes #35020

Signed-off-by: Keith Mattix II <keithmattix@microsoft.com>