xds: refactor xds-resources-delegate and move it to xds-manager (#38495)
Commit Message: xds: refactor xds-resources-delegate and move it to
xds-manager

This PR moves the xDS-Resources-Delegate from the cluster-manager to the
xds-manager.
Note that in this PR we add a getter for the xDS-Resources-Delegate, but
that will be removed in the future when the rest of the xDS-related
components are moved from the cluster-manager to the xds-manager.


Signed-off-by: Adi Suissa-Peleg <adip@google.com>