When building with gperftools always link the library
There is a known issue in gperftools - when linking statically profiling
triggered by environment variables does not work, because linker does
not see the right profiling functions used and eliminates them
completely.

I'm not entirely sure why the issues has not been fixed after all this
time, but it's relatively well documented (see
https://github.com/gperftools/gperftools/issues/961,
https://github.com/gperftools/gperftools/issues/415, etc).

There are a few ways to address the issues:

1. Link dynamically, this way linker would not have enough information
   to eliminate "unused" functions, because it cannot actually tell if
   they are trully unused
2. Explicitly tell linker that some functions must be preserved using
   -Wl,--undefined=<function name>
3. Use --whole-archive to prevent linker from eliminating functions it
   deems "unused".

This change goes for the third option as the least fragile,
linkalways=True, when it works, instructs linker to use --whole-archive
on the library marked with linkalways=True.

NOTE: linkalways=True does not work on cc_library when the source is a
static archive (i.e., *.a file instead of *.h, *.cc), but to my surprise
it works as expected in configure_make.

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>