Jwks fetcher use scheme from uri (#37978)
<!--
!!!ATTENTION!!!

If you are fixing *any* crash or *any* potential security issue, *do
not*
open a pull request in this repo. Please report the issue via emailing
envoy-security@googlegroups.com where the issue will be triaged
appropriately.
Thank you in advance for helping to keep Envoy secure.

!!!ATTENTION!!!

For an explanation of how to fill out the fields, please see the
relevant section
in
[PULL_REQUESTS.md](https://github.com/envoyproxy/envoy/blob/main/PULL_REQUESTS.md)
-->

Commit Message:
jwks_fetcher: set :scheme header from uri

Additional Description:
the jwks fetcher in the jwt_authn filter does not set the `:scheme`
pseudo header according to the configured uri. Which leads to the header
value always being set to 'http'. This might be problematic if a jwks
service (using http/2 or higher) is expecting this header to be set to
'https' if the transport is TLS.

The same issue exists probably with other code using the `async_client`
code to create similar 'out of band' requests (e.g requests to the
authorization server from the oauth2 filter)

I've opened another PR #37588 which tried to 'fix' this issue on a lower
level, setting the `:scheme` header according to the upstream. But due
to risk avoidance I now try to fix it on the level of the request
performed (according to the configured uri) and only for the
jwks_fetcher. If this approach is accepted other `ascync_client` usages
may be fixed the same way.

I know there are no jwks_fetcher tests in this PR at the moment. I first
wanted to check if this approach is acceptable before I continue to
spend time and effort for it.

Risk Level:
low

Testing:
unit-tests added

Docs Changes:
-

Release Notes:
see current.yaml

Platform Specific Features:
-

[Optional Runtime guard:]
[Optional Fixes #Issue]
[Optional Fixes commit #PR or SHA]
[Optional Deprecated:]
[Optional [API
Considerations](https://github.com/envoyproxy/envoy/blob/main/api/review_checklist.md):]

---------

Signed-off-by: Frank Meier <frank.meier@ergon.ch>