Stop compiler from inlining releaseAssertInAFunction in test
It looks that assert tests were recently updated to validate that
triggered asserts print a stack trace. However stacktraces that are
produced depend on the compilation parameters and compiler, i.e., if
compiler will decide to inline releaseAssertInAFunction it will not show
up in the stack trace breaking the test, even though nothing is actually
wrong.

That's exactly what I noticed when building Envoy using clang-18 as we
are trying to move Envoy CI to clang-18.

This change adds an attribute to the releaseAssertInAFunction that would
prevent commonly used compilers (and importantly clang specifically)
from inlining the function and breaking the test.

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>

Signed-off-by: Ryan Northey <ryan@synca.io>