Update v8 patch to fix ASAN issue detected by clang-18
This is a workaround to help us migrate to clang-18 for Envoy CI.
clang-18 detected a "use-after-free" type of issue in v8 code and that
fails Envoy CI ASAN runs.

This issue was fixed upstream already (see
https://github.com/v8/v8/commit/df459f0e7eda0d1c50698284aab6f424057e076a)
and updating v8 should resolve the problem. However for now I'm working
around that with a patch update because:

1. Updating v8 will fix this problem, but might also introduce new
   issues, so I'd like to finish with clang-18 migration for now and
   update v8 once clang-18 migration is settled
2. v8 is a bit of special dependency, as Envoy does not use upstream v8
   directly but instead relies on some kind of processed version, that's
   probably one of the reasons why we don't update v8 as often.

NOTE: I filed a bug to consider streamlining v8 updates to work on it
once we are done with clang-18 migration:
https://github.com/envoyproxy/envoy/issues/38609

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>