Set RCD for 100% drop_overload case as unconditional_drop_overload (#38408)
This is a follow up commit of drop_overload
support:https://github.com/envoyproxy/envoy/pull/30827.

It is observed that in some situations, EDS sends drop_overload ratio as
100% for some clusters do not have healthy endpoints in the cluster. For
such cases, it is preferred that the RCD of the local reply set to
"unconditional drop overload" to make it a little bit different from the
normal "drop overload" case.

---------

Signed-off-by: Yanjun Xiang <yanjunxiang@google.com>