docs: correct route check docs (#38611)
Commit Message:

Fixes the docs for the route checker tool which do not currently contain
a valid test case.

The method parameter is marked as required yet the description implies
it is not. From experimentation it does appear to be required, but I
have not dived deep enough into the source code to fully verify this.

Signed-off-by: Ho-Yon Mak <ho-yon.mak@multiverse.io>