Workaround coverage issues observed on clang-18 (#38628)
I described the problem with coverage and possible solutions to it in
https://github.com/envoyproxy/envoy/issues/37911#issuecomment-2659837724.

TL;DR: clang source-based coverage (which is what we currently use) is
subtly broken when dynamic linking is involved (which is what we
currently do for coverage tests).

The issue shows itself as lack of coverage of some functions defined in
a header file (class members, free-standing inline functions or function
temapltes), even though there are tests that cover it.

The issue is a result of a bug in LLVM/Clang/coverage tooling (it's
debatable where exactly is the bug fix should be):
https://github.com/llvm/llvm-project/issues/32849.

The issue symptoms are not very predictable, and can be affected by a
bunch of factors starting from the order of linking and ending with
optimization level of the compiler.

Migrating to clang-18 changed enough things that we started to see the
issue, even though, from all I know the issue is not specific to
clang-18 and exists even in the currently used by Envoy CI clang-14.

There are a few ways to work around/fix the issue, that could be
considered:

1. Stop using dynamic linking - this will resolve the issue completely
for Envoy codebase, but at the cost of potentially introducing other
issues (hustorically Envoy coverage tests didn't use dynamic linking,
but we switched to it to deal with some issues)
2. Force Clang to emit code for each function definition it sees even if
it does not see it being used, either globally via `-femit-all-decls`
compiler flag or locally, on per function basis, marking functions with
`used` attribute.
3. Move affected functions from the header files to cc files.

NOTE: There is no upstream fix for the issue in Clang/LLVM, so I'm not
mentioning it as an option at the moment.

This change uses a combination of:

1. moving functions from header files to cc files [preferred]
2. marking functions with `used` attribute [when the above is not
desirable]

The reason why I didn't go for more generic options are as follows:

1. It's good to move function definitions that don't have to be inlined
to a cc file, it reduces build-time dependencies and the amount of work
compiler needs to do;
2. The code does not build with `-femig-all-decls` and the build issues
are significant enough, so that we cannot just patch our way out; plus
it undoes a compiler build-time optimization (e.g., we will basically
force the compiler to do work that might not be necessary)
3. We currently cannot use static linking for coverage tests, because it
does not work due to relocation overflows (it could be addressed by
migrating to clang-18, but on clang-14 it just does not work) and the
effects on the Envoy CI are not clear (remember that dynamic linking was
introduced to address other issues and we don't want those to return).

The only place where I used `used` attribute instead of moving function
definition to a header file was in WASM code where the same code
apparently was compiled both into Envoy and into a WASM binary, so I
figured we can tolerate a `used` attribute in that case, given how
uncommon that setup is. Though, I don't feel particularly strongly about
it and can change things.

So this change does not address the issue with coverage completealy, but
it improves coverage results on clang-18 significantly enough, where it
would not be a concern and we can switch to clang-18. Once we are on
clang-18 we are not only going to enjoy more extensive coverage results,
but will be in position to consider static linking for coverage tests
again to fix the issue permanently for us.

Related to https://github.com/envoyproxy/envoy/issues/37911 and fixes
one of the issues that block clang-18 adoption in Envoy CI (specifically
addresses one of the failures in
https://github.com/envoyproxy/envoy/pull/38571).

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>