cds: clarify misleading/incomplete `info` logs
Lower the log message that counts inputs to `onConfigUpdate` to
`debug`. Because it's at `info`, this output appears in Envoy log output
by default, and it's confusing in that context. It is just counting
clusters returned by CDS (not necessary an actual addition over what
already exists) and existing clusters not returned by CDS (not
necessarily a reduction! any statically configured clusters will be
counted as "removed" here even though they will definitely not be
actually removed by this function). The changes actually applied by the
CDS update are logged at the end of the function.

I think semantically what's being logged in this first statement is
`debug`-level stuff: it's really just logging some information about
function arguments, not information useful to end users.

And improve the more useful log message to include a count of actually
removed clusters, just as it counts added/updated and skipped ones.

Signed-off-by: Ian Kerins <git@isk.haus>