Refactor #ifdefs in stats.cc to help coverage tools figure out dead code (#38672)
Commit Message:

configureBackgroundMemoryRelease() method is only called in tcmalloc
builds. Coverage reports are built with gperftools tcmalloc instead of
the regular one.

This creates a situation when coverage report does not recognize that
configureBackgroundMemoryRelease cannot be called at all and declares it
as not covered pushing the coverage stats down.

Given that configureBackgroundMemoryRelease currently cannot be covered
by tests in principle, I think it makes sense to exclude the dead parts
of the code from the coverage reports.

This refactoring moves the logic from AllocatorManager::AllocatorManager
to configureBackgroundMemoryRelease (including all the #ifdefs), and
instead of calling configureBackgroundMemoryRelease conditionally calls
it unconditionally.

This way all the parts of the configureBackgroundMemoryRelease that
cannot be executed with coveraged, will not be compiled in the final
binary and will not show up as missing coverage in the coverage report.

NOTE: We are working on changing Envoy CI to clang-18, and while
clang-18 generally covers more lines (because it supports coverage for
more C++ constructs than currently used clang-14), it also causes enough
changes that we fell below the coverage threshold in
source/common/memory.cc. This change will help us to get over that,
without changing coverage thresholds.

Additional Description:

Related to https://github.com/envoyproxy/envoy/issues/37911 and fixes
one of the issues that block clang-18 adoption in Envoy CI (specifically
addresses one of the failures in
https://github.com/envoyproxy/envoy/pull/38571).

Risk Level: low (non-functional change)
Testing: running coverage on clang-18
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

+cc @phlax

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>