bootstrap: allow to fail listener health check at startup
This PR allows to start healthcheck filter with failed status.

In some setup, Envoy acts as backend for a downstream proxy.
At startup, it is useful to wait for the xDS config coming from the
control plane, and run some readiness checks, before enabling the
healthcheck.
This allows to be in a better control when we can actually allow traffic
to be received.

For example we face races where the listeners becomes available but part
of our pod setup is not yet ready. We therefore need more control to be
able to decide when we allow traffic to come in.

Signed-off-by: William Dauchy <william.dauchy@datadoghq.com>