formatter: make the CEL and METADATA as built in formatter (#38740)
Commit Message: formatter: make the CEL and METADATA as built in
formatter
Additional Description:

This make the CEL and METADATA as built in formatter and could be used
everywhere. Now, we needn't to configure it. After this PR, it's
possible to use CEL at the header mutation which would be helpful.

Risk Level: low.
Testing: unit/integration.
Docs Changes: added.
Release Notes: added.
Platform Specific Features: n/a.

---------

Signed-off-by: wangbaiping(wbpcode) <wbphub@gmail.com>