Make balsa parser not reject newlines between requests (#38732)
Commit Message: Make balsa parser not reject newlines between requests
Additional Description: Per
https://github.com/envoyproxy/envoy/issues/38695, rejecting newlines
between requests is in violation of a "SHOULD" in the RFCs. Paul
suggested removing this check as a possible fix for the behavior, which
I believe makes sense because the behavior is the same without this
extra check, in case of an actual invalid method, just slightly slower.
An optimization on an error path that costs an extra check on the common
path is likely to be overall a performance detriment, so removing it
will: 1. simplify the code, 2. fix the behavior to what "SHOULD" happen,
and 3. imperceptibly improve performance.
Risk Level: Very low. A behavior change only replacing errors with
non-errors (correctly), with no change to normal handling.
Testing: Correct behavior and correct error behavior covered by existing
tests. Added a regression test for correct behavior specifically around
newlines.
Docs Changes: n/a
Release Notes: Yes.
Platform Specific Features: n/a

---------

Signed-off-by: Raven Black <ravenblack@dropbox.com>